Move mirror configuration to specific OCI implementation
