+++ internal/oci/oci.go
@@ -7,6 +7,10 @@ import (
 	"github.com/opencontainers/go-digest"
 )
 
+type UnknownDocument struct {
+	MediaType string `json:"mediaType,omitempty"`
+}
+
 type Client interface {
 	Subscribe(ctx context.Context) (<-chan Image, <-chan error)
 	ListImages(ctx context.Context) ([]Image, error)
@@ -15,8 +19,4 @@ type Client interface {
 	GetSize(ctx context.Context, dgst digest.Digest) (int64, error)
 	WriteBlob(ctx context.Context, dst io.Writer, dgst digest.Digest) error
 	GetBlob(ctx context.Context, dgst digest.Digest) ([]byte, string, error)
-}
-
-type UnknownDocument struct {
-	MediaType string `json:"mediaType,omitempty"`
 }
